build/MSVC: Require Visual Studio 2019 16.8.x
We need to use the newer and more C99/C11 conformant preprocessor that is
activated by /Zc:preprocessor, so that we can try to clean up the SIMD code
that is eventually to be shared between Visual Studio and GCC/CLang.

Split up the compiler flags to check for to make things a bit more clear.